separated matrix operations into files
